# Docs - contract/openapi.yaml
openapi: 3.0.1
info:
  title: Homely API
  description: HTTP JSON dùng {statusCode,message,data}; DTO trong data dùng snake_case, VND là chuỗi số nguyên. Thanh toán và hoàn tiền là sandbox.
  version: 1.6.0
servers:
- description: Local backend; paths include /api/v1
  url: http://localhost:8080
paths:
...
  /api/v1/wishlist/{roomId}:
    post:
      tags:
      - wishlist-controller
      summary: Save room to wishlist
      operationId: saveRoom
      parameters:
      - name: roomId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: Idempotency-Key
        in: header
        description: Giữ nguyên khi retry cùng thao tác; cùng khóa nhưng payload khác trả 409.
        required: true
        schema:
          maxLength: 128
          minLength: 1
          type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                required:
                - statusCode
                - message
                - data
                type: object
                properties:
                  statusCode:
                    type: integer
                    format: int32
                    example: 200
                  message:
                    type: string
                  data:
                    type: object
                    additionalProperties:
                      type: string
        default:
          description: HTTP status giữ nguyên; data chứa ProblemDTO với code, field_errors và trace_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestResponseProblemDTO'
      security:
      - bearerAuth: [
          ]
    delete:
      tags:
      - wishlist-controller
      summary: Remove room from wishlist
      operationId: unsaveRoom
      parameters:
      - name: roomId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                required:
                - statusCode
                - message
                - data
                type: object
                properties:
                  statusCode:
                    type: integer
                    format: int32
                    example: 200
                  message:
                    type: string
                  data:
                    type: object
                    additionalProperties:
                      type: string
        default:
          description: HTTP status giữ nguyên; data chứa ProblemDTO với code, field_errors và trace_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestResponseProblemDTO'
      security:
      - bearerAuth: [
          ]
...
  /api/v1/reports:
    post:
      tags:
      - report-controller
      summary: Create report
      operationId: createReport
      parameters:
      - name: Idempotency-Key
        in: header
        description: Giữ nguyên khi retry cùng thao tác; cùng khóa nhưng payload khác trả 409.
        required: true
        schema:
          maxLength: 128
          minLength: 1
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportCreateRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                required:
                - statusCode
                - message
                - data
                type: object
                properties:
                  statusCode:
                    type: integer
                    format: int32
                    example: 200
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/ReportDTO'
        default:
          description: HTTP status giữ nguyên; data chứa ProblemDTO với code, field_errors và trace_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestResponseProblemDTO'
      security:
      - bearerAuth: [
          ]
...
  /api/v1/wishlist:
    get:
      tags:
      - wishlist-controller
      summary: List saved rooms
      operationId: getSavedRooms
      parameters:
      - name: pageable
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                required:
                - statusCode
                - message
                - data
                type: object
                properties:
                  statusCode:
                    type: integer
                    format: int32
                    example: 200
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/PageResponseMapStringObject'
        default:
          description: HTTP status giữ nguyên; data chứa ProblemDTO với code, field_errors và trace_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestResponseProblemDTO'
      security:
      - bearerAuth: [
          ]
  /api/v1/wishlist/{roomId}/check:
    get:
      tags:
      - wishlist-controller
      summary: Check if room is saved
      operationId: checkSaved
      parameters:
      - name: roomId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                required:
                - statusCode
                - message
                - data
                type: object
                properties:
                  statusCode:
                    type: integer
                    format: int32
                    example: 200
                  message:
                    type: string
                  data:
                    type: object
                    additionalProperties:
                      type: boolean
        default:
          description: HTTP status giữ nguyên; data chứa ProblemDTO với code, field_errors và trace_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestResponseProblemDTO'
      security:
      - bearerAuth: [
          ]
...
  /api/v1/reports/my:
    get:
      tags:
      - report-controller
      summary: Get my reports
      operationId: getMyReports
      parameters:
      - name: pageable
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                required:
                - statusCode
                - message
                - data
                type: object
                properties:
                  statusCode:
                    type: integer
                    format: int32
                    example: 200
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/PageResponseReportDTO'
        default:
          description: HTTP status giữ nguyên; data chứa ProblemDTO với code, field_errors và trace_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestResponseProblemDTO'
      security:
      - bearerAuth: [
          ]
...
  /api/v1/amenities:
    get:
      tags:
      - amenity-controller
      summary: List all amenities
      operationId: getAllAmenities
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                required:
                - statusCode
                - message
                - data
                type: object
                properties:
                  statusCode:
                    type: integer
                    format: int32
                    example: 200
                  message:
                    type: string
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Amenity'
        default:
          description: HTTP status giữ nguyên; data chứa ProblemDTO với code, field_errors và trace_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestResponseProblemDTO'
  /api/v1/amenities/{id}:
    get:
      tags:
      - amenity-controller
      summary: Get amenity detail
      operationId: getAmenity
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                required:
                - statusCode
                - message
                - data
                type: object
                properties:
                  statusCode:
                    type: integer
                    format: int32
                    example: 200
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/Amenity'
        default:
          description: HTTP status giữ nguyên; data chứa ProblemDTO với code, field_errors và trace_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestResponseProblemDTO'
...
components:
  schemas:
...
    ReportCreateRequest:
      required:
      - reasonCode
      - targetId
      - targetType
      type: object
      properties:
        targetType:
          maxLength: 50
          minLength: 0
          type: string
        targetId:
          type: integer
          format: int64
        reasonCode:
          maxLength: 50
          minLength: 0
          type: string
        description:
          maxLength: 2000
          minLength: 0
          type: string
    ReportDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        description:
          type: string
        status:
          type: string
        reporter_id:
          type: integer
          format: int64
        target_type:
          type: string
        target_id:
          type: integer
          format: int64
        reason_code:
          type: string
        created_at:
          type: string
          format: date-time
...
    Pageable:
      type: object
      properties:
        page:
          minimum: 0
          type: integer
          format: int32
        size:
          minimum: 1
          type: integer
          format: int32
        sort:
          type: array
          items:
            type: string
    PageResponseMapStringObject:
      type: object
      properties:
        data:
          type: array
          items:
            type: object
            additionalProperties:
              type: object
        total_elements:
          type: integer
          format: int64
        total_pages:
          type: integer
          format: int32
        current_page:
          type: integer
          format: int32
        page_size:
          type: integer
          format: int32
...
    PageResponseReportDTO:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ReportDTO'
        total_elements:
          type: integer
          format: int64
        total_pages:
          type: integer
          format: int32
        current_page:
          type: integer
          format: int32
        page_size:
          type: integer
          format: int32
...
    Amenity:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        icon:
          type: string
        category:
          type: string
...
    FieldErrorDTO:
      type: object
      properties:
        field:
          type: string
        reason:
          type: string
    ProblemDTO:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
        code:
          type: string
        timestamp:
          type: string
          format: date-time
        field_errors:
          type: array
          items:
            $ref: '#/components/schemas/FieldErrorDTO'
        trace_id:
          type: string
    RestResponseProblemDTO:
      required:
      - statusCode
      - message
      - data
      type: object
      properties:
        statusCode:
          type: integer
          format: int32
          example: 500
        message:
          type: string
        data:
          $ref: '#/components/schemas/ProblemDTO'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...